Removed string variable str.
It was not adding clarity to the code. It was used always in the same
way: read a string, then parse it. We do not need an intermediate
variable for it.

The usual reason for introducing a new variable is either:

a) To make the code clearer. This happens when the variable's name
  refers to an important concept like 'euroToPoundsRatio',
  'databaseConnection', or any other partial result.

b) To make the code faster. This happens when the variable holds the
  result of some intermediate computation that is expensive and you
  should not perform the same computation several times but just once,
  e.g. fixed-rate interest payment calculations over several months,
  word-counting over the same text, reading the same web page from a
  network connection, etc.

Neither of these conditions is true in this case, so I delete the
variable to make the code shorter (shorter code is clearer, provided
all other factors are equal).