refactoring: replaced System.console().readLine() by use of java.util.Scanner
Rationale:
System.console() returns null on IDEs like IntelliJ and Eclipse, so
the old version would crash the program with a NullPointerException
when executed on one of these platforms. In contrast, reading via
java.util.Scanner is more robust since it does not require an
explicit Console object and works on the above IDEs as well.